[#28] Refactoring AccountRolesConverter 테스트 코드 위치 변경   - 공통적으로 필요한 부분이라, 의견 조율 후 domin 패키지로 이동
